intentsearch-app-main/src/lambda/NaturalLanguage/requirements.txt
google-genai
requests
aws-lambda-powertools
aws-xray-sdk
amazon-dax-client
launchdarkly-server-sdk
cachetools
urllib3
inflect
